+++ Models/ProjectTask.cs
@@ -9,13 +9,13 @@ public class ProjectTask
 {
     [Key]
     [Column("project_task_id")]
-    public long ProjectTaskId { get; set; }
+    public int ProjectTaskId { get; set; }
 
     [Column("project_task_phase_id")]
     public int ProjectTaskPhaseId { get; set; }
 
     [Column("project_header_id")]
-    public long ProjectHeaderId { get; set; }
+    public int ProjectHeaderId { get; set; }
 
     [Column("task_no")]
     [MaxLength(25)]
