+++ Services/ITaskNotificationService.cs
@@ -8,6 +8,6 @@ public interface ITaskNotificationService
     Task<List<PendingTaskDto>> GetPendingTasksAsync(string userId);
     Task<List<TaskItem>> GetPendingTaskItemsAsync(string userId);
     Task<string> GenerateTaskSummaryMessageAsync(List<PendingTaskDto> tasks, string userName);
-    Task<bool> SaveTaskTrackingAsync(string userId, string lineUserId, TimesheetEntry entry, long projectTaskId);
+    Task<bool> SaveTaskTrackingAsync(string userId, string lineUserId, TimesheetEntry entry, int projectTaskId);
     Task<List<string>> GetIssueTypesAsync();
 }
